fixed graph not showing bug
